+++ meteor-server/src/LoginMethodResponse.js
@@ -1,0 +1,11 @@
+import {Meteor} from 'meteor/meteor'
+
+export default function (options) {
+  return {
+    LoginMethodResponse: {
+      user ({id}) {
+        return Meteor.users.findOne(id)
+      }
+    }
+  }
+}